+++ CHANGELOG.md
@@ -14,8 +14,9 @@
 - Added dot and glow dot optional accents for unread messages
 - Added window size persistence - [#217](https://github.com/hkdb/aerion/issues/217)
 - Added mini calendar for easier navigation - [#368](https://github.com/hkdb/aerion/issues/368)
-- Fixed post delete ghost messages on message list
+- Added Empty Spam button for spam folder - [#239](https://github.com/hkdb/aerion/issues/239)
 - New app icon with full hicolor size set + scalable SVG - [#395](https://github.com/hkdb/aerion/issues/395) (Special thanks to @gianlucamazza) - [#99](https://github.com/hkdb/aerion/issues/99) 
+- Fixed post delete ghost messages on message list
 - Fixed thread sender and preview in message list - [#169](https://github.com/hkdb/aerion/issues/169)
 - Fixed comma display name - [#398](https://github.com/hkdb/aerion/issues/398)
 - Fixed composer draft format detection - [#420](https://github.com/hkdb/aerion/issues/420)